+++ src/shapes/mod.rs
@@ -1,5 +1,6 @@
 pub mod collection;
 pub mod rendering;
+pub mod collection_mutation;
 
 use crate::drawable::{DrawType, Drawable};
 use graphics_shapes::circle::Circle;